requirements.txt
django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
django-filter==22.1
mysqlclient==2.1.1